[feat] 키워드, 요약 부분 뷰모델 파일 분리
